\name{apoe}
\docType{data}
\alias{apoe}
\alias{apoe7}
\title{Genetic data of APOE gene region}
\description{
 This data set came from a re-sequenced data of APOE gene region in the Molecular Diversity and Epidemiology of Common Disease (MDECODE) database. Sixteen polymorphic sites were included. "apoe7" data contains the genetic data of seven single nucleotide polymorphisms with allele frequencies higher than 0.1 from the apoe data.
}
\usage{data(apoe)}
\format{A matrix with 48 rows and 32 columns}
\source{http://droog.gs.washington.edu/mdecode/}
\references{
Nickerson, D. A., S. L. Taylor, S. M. Fullerton, K. M. Weiss, A. G. Clark et al. (2000) Sequence diversity and large-scale typing of SNPs in the human apolipoprotein E gene. \emph{Genome Res} 10: 1532-1545.
}
\keyword{datasets}
