\name{apoeG}
\docType{data}
\alias{apoeG}
\alias{genoT}
\alias{genoC}
\title{Sequencing data of APOE gene region from the 1000 Genomes Project}
\description{
 This data set came from a re-sequenced data of APOE gene region from the 1000 Genomes Project. Thirty three polymorphic sites with allele frequencies higher than 0.001 were included for the original data set, apoeG. The test data sets, apoeT and apoeC, indicate the data of 100 controls and 100 cases respectively when the dominant variant is 15th variant with the odds ratio of 3.
}
\usage{data(apoeG)}
\format{A matrix with 33 rows and 2184 columns}
\source{ftp://ftp-trace.ncbi.nih.gov/1000genomes/ftp/release/20110521/}
\references{
Abecasis, G. R. et al. (2010) A map of human genome variation from population-scale sequencing. \emph{Nature} 467, 1061-1073.
}
\keyword{datasets}
